+++ inst/htmlwidgets/lib/geoparquet-wasm/geoparquet2arrow.js
@@ -1,6 +1,6 @@
 import * as geoparquet from "https://cdn.jsdelivr.net/npm/@geoarrow/geoparquet-wasm/esm/+esm";
 
-async function parquet2arrow(pq) {
+async function geoparquet2arrow(pq) {
   await geoparquet.default();
   const parquetBytes = new Uint8Array(await pq.arrayBuffer());
   const wasmTable = geoparquet.readGeoParquet(parquetBytes);
@@ -8,4 +8,4 @@ async function parquet2arrow(pq) {
   return arrow_table;
 }
 
-Object.assign(window, {parquet2arrow});
+Object.assign(window, {geoparquet2arrow});