+++ inst/tinytest/test-extJSLibs.R
@@ -1,6 +1,6 @@
 # File created by roxut; edit the function definition file, not this file
 
-# Test found in extJSLibs.R:18 (file:line)
+# Test found in extJSLibs.R:23 (file:line)
   
-expect_length(extJSLibs(), 3)
-expect_length(names(extJSLibs()), 3)
+expect_length(extJSLibs(), 4)
+expect_length(names(extJSLibs()), 4)